requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
PyQt6>=6.6.0
google-generativeai>=0.3.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
python-dotenv>=1.0.0

